+++ src/test/java/com/exclamationlabs/connid/box/testutil/LocalBoxConnector.java
@@ -1,0 +1,16 @@
+package com.exclamationlabs.connid.box.testutil;
+
+import com.exclamationlabs.connid.box.BoxConnector;
+
+/**
+ * BoxConnector implementation for local testing environment which uses mock Box API instead of the real Box API.
+ *
+ * @author Hiroyuki Wada
+ */
+public class LocalBoxConnector extends BoxConnector {
+
+    @Override
+    protected void authenticateResource() {
+        boxAPI = MockBoxAPIHelper.instance().getAPIConnection();
+    }
+}
